% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_pie_glyph.R
\name{draw_key_pie}
\alias{draw_key_pie}
\title{Legend key for the pie glyphs}
\usage{
draw_key_pie(data, params, size)
}
\arguments{
\item{data}{A single row data frame containing the scaled aesthetics to
display in this key}

\item{params}{A list of additional parameters supplied to the geom.}

\item{size}{Width and height of key in mm.}
}
\value{
A grid grob
}
\description{
Controls the aesthetics of the legend entries for the pie glyphs
}
\seealso{
\code{\link[ggplot2:draw_key]{draw_key}}
}
