% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_utils_grid.R
\name{upgradeUnit.unit.list}
\alias{upgradeUnit.unit.list}
\title{Upgrading units for a list}
\usage{
\method{upgradeUnit}{unit.list}(x)
}
\arguments{
\item{x}{A \code{unit.list} object.}
}
\value{
A combined unit object.
}
\description{
Upgrading units for a list
}
