% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_pie_glyph.R
\name{scale_radius_discrete}
\alias{scale_radius_discrete}
\alias{scale_radius_manual}
\alias{scale_radius_continuous}
\alias{scale_radius}
\title{Scales for the pie radius}
\usage{
scale_radius_discrete(..., range = c(0.5, 1.5), unit = "cm")

scale_radius_manual(..., values, unit = "cm", breaks = waiver(), na.value = NA)

scale_radius_continuous(..., range = c(0.5, 1.5), unit = "cm")

scale_radius(..., range = c(0.5, 1.5), unit = "cm")
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[ggplot2:continuous_scale]{continuous_scale}}
  \describe{
    \item{\code{minor_breaks}}{One of:
\itemize{
\item \code{NULL} for no minor breaks
\item \code{waiver()} for the default breaks (one minor break between
each major break)
\item A numeric vector of positions
\item A function that given the limits returns a vector of minor breaks. Also
accepts rlang \link[rlang:as_function]{lambda} function notation.
}}
    \item{\code{oob}}{One of:
\itemize{
\item Function that handles limits outside of the scale limits
(out of bounds). Also accepts rlang \link[rlang:as_function]{lambda}
function notation.
\item The default (\code{\link[scales:oob]{scales::censor()}}) replaces out of
bounds values with \code{NA}.
\item \code{\link[scales:oob]{scales::squish()}} for squishing out of bounds values into range.
\item \code{\link[scales:oob]{scales::squish_infinite()}} for squishing infinite values into range.
}}
    \item{\code{na.value}}{Missing values will be replaced with this value.}
    \item{\code{expand}}{For position scales, a vector of range expansion constants used to add some
padding around the data to ensure that they are placed some distance
away from the axes. Use the convenience function \code{\link[ggplot2:expansion]{expansion()}}
to generate the values for the \code{expand} argument. The defaults are to
expand the scale by 5\% on each side for continuous variables, and by
0.6 units on each side for discrete variables.}
    \item{\code{position}}{For position scales, The position of the axis.
\code{left} or \code{right} for y axes, \code{top} or \code{bottom} for x axes.}
    \item{\code{super}}{The super class to use for the constructed scale}
  }}

\item{range}{a numeric vector of length 2 that specifies the minimum and
maximum size of the plotting symbol after transformation.}

\item{unit}{Unit for the radius of the pies. Default is 'cm', but other units like 'in', 'mm', etc. can be used.}

\item{values}{a set of aesthetic values to map data values to. The values
will be matched in order (usually alphabetical) with the limits of the
scale, or with \code{breaks} if provided. If this is a named vector, then the
values will be matched based on the names instead. Data values that don't
match will be given \code{na.value}.}

\item{breaks}{One of:
\itemize{
\item \code{NULL} for no breaks
\item \code{waiver()} for the default breaks computed by the
\link[scales:trans_new]{transformation object}
\item A numeric vector of positions
\item A function that takes the limits as input and returns breaks
as output (e.g., a function returned by \code{\link[scales:breaks_extended]{scales::extended_breaks()}}).
Also accepts rlang \link[rlang:as_function]{lambda} function notation.
}}

\item{na.value}{The aesthetic value to use for missing (\code{NA}) values}
}
\description{
\code{scale_radius_*()} is useful for adjusting the radius of the pie glyphs.
}
\examples{
library(PieGlyph)
library(DImodels)
library(tidyr)
library(dplyr)
library(ggplot2)

## Load the data
data(sim1)
sim1$Evenness <- DI_data_E_AV(prop = 3:6, data = sim1)$E

## Convert data into long-format
plot_data <- sim1 \%>\% filter(block == 1) \%>\%
                      pivot_longer(cols = paste0('p',1:4),
                                   names_to = 'Species', values_to = 'Prop')

## Create plot
p <- ggplot(data = plot_data)+
    geom_pie_glyph(aes(x = community, y = response, radius = Evenness),
    categories = 'Species', values = 'Prop', colour = NA)+
    labs(y = 'Response', x = 'Community')+
    theme_classic()

p + scale_radius_continuous(range = c(1, 2))

q <- ggplot(data = plot_data)+
    geom_pie_glyph(aes(x = community, y = response,
                       radius = as.factor(Evenness)),
                 categories = 'Species', values = 'Prop', colour = 'black')+
    labs(y = 'Response', x = 'Community')+
    theme_classic()

q + scale_radius_discrete(range = c(0.1, 0.2), unit = 'in',
                          name = 'Evenness')

q + scale_radius_manual(values = c(5, 20, 10, 15), unit = 'mm',
                        labels = LETTERS[1:4], name = 'E')
}
